src/consumer_influxdb/requirements.txt
kafka-python
influxdb3-python
pandas